added to price max digits 6 and decimal_places 2

+++ shop/models.py
@@ -3,7 +3,7 @@ from django.db import models
 # Create your models here.
 class Products(models.Model):
     title = models.CharField(max_length=250)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=6, decimal_places=2)
     discount_price = models.FloatField()
     category = models.CharField(max_length=250)
     description = models.TextField()
